+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Warnings now have more comprehensive descriptions and are sorted by severity
+- `DumpIsUsed`: Report only `dump` calls with non-literal arguments
 
 ## [0.3.1] - 2024-09-24